% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5_view.R
\name{a5_viewOutput}
\alias{a5_viewOutput}
\title{Shiny output binding for a5_view}
\usage{
a5_viewOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output variable name.}

\item{width, height}{Widget dimensions.}
}
\description{
Shiny output binding for a5_view
}
