% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5_view.R
\name{renderA5_view}
\alias{renderA5_view}
\title{Shiny render function for a5_view}
\usage{
renderA5_view(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns an a5_view widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression?}
}
\description{
Shiny render function for a5_view
}
